Move private VC creation/deletion handlers
