+++ model/private-vc.js
@@ -1,4 +1,5 @@
 const db = require('./db');
+const Guild = require('./guild');
 
 const PrivateVC = {
     /** {Object} */
@@ -59,6 +60,50 @@ const PrivateVC = {
     getPrivateChannelsList: () => {
         return Array.from(new Set(Object.values(PrivateVC.list)));
     },
+
+    /**
+     * @param {VoiceState} oldVoiceState
+     */
+    privateVoiceChatRequestHandler: async (oldVoiceState) => {
+        const member = oldVoiceState.member;
+        await Promise.all([
+            member.guild.channels.create(`[Private] ${member.displayName}`, {
+                type: 'voice',
+                parent: Guild.smallVoiceCategoryChannel,
+            }),
+            member.guild.channels.create("[Waiting room] ⬆️", {
+                type: 'voice',
+                parent: Guild.smallVoiceCategoryChannel,
+            })
+        ]).then(async ([privateChannel, waitingRoomChannel]) => {
+            await member.voice.setChannel(privateChannel);
+            return PrivateVC.add(privateChannel.id, member.id).catch(exception => {
+                exception.payload = [ privateChannel, waitingRoomChannel ];
+                throw exception;
+            });
+        }).catch(async (exception) => {
+            Logger.exception(exception);
+            await Guild.botChannel.send(trans('model.privateVC.errors.creationFailed.mods', [member.toString()]));
+            await member.send(trans('model.privateVC.errors.creationFailed.member'));
+            await member.voice.setChannel(oldVoiceState.channel);
+            await Promise.all(exception.payload.map(channel => channel.delete()));
+        });
+    },
+
+    /**
+     * @param {VoiceState} oldVoiceState
+     */
+    privateVoiceChatDeletionHandler: async (oldVoiceState) => {
+        const member = oldVoiceState.member;
+        const privateChannel = oldVoiceState.channel;
+        const waitingRoomChannel = Guild.discordGuild.channels.cache.find(channel => channel.rawPosition === privateChannel.rawPosition + 1);
+        await Promise.all([privateChannel.delete(), waitingRoomChannel.delete()]).then(async () => {
+            return PrivateVC.remove(member.id).catch(async exception => {
+                Logger.exception(exception);
+                await Guild.botChannel.send(trans('model.privateVC.errors.deletionFailed', [privateChannel.id, member.toString()]));
+            });
+        });
+    }
 }
 
 module.exports = PrivateVC;
